+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0 (2022-04-23)
+  * rename `unWithRefine` to `withoutRefine`
+  * reorder `reallyUnsafeEnforce` typevars for visible type application
+    convenience
+
 ## 0.1.0 (2022-04-22)
 Initial release.